arm64: RME: Allocate/free RECs to match vCPUs
The RMM maintains a data structure known as the Realm Execution Context
(or REC). It is similar to struct kvm_vcpu and tracks the state of the
virtual CPUs. KVM must delegate memory and request the structures are
created when vCPUs are created, and suitably tear down on destruction.

See Realm Management Monitor specification (DEN0137) for more information:
https://developer.arm.com/documentation/den0137/

Change-Id: I87b8e5f12fcf2a0244ce2d2f8a213871a8fee717
Signed-off-by: Steven Price <steven.price@arm.com>